+++ src/patterns/composite/VehicleAssembly.cpp
@@ -8,8 +8,8 @@ void VehicleAssembly::removePart(VehiclePart* part) {
     // Implementation of removal is omitted for simplicity
 }
 
-double VehicleAssembly::getCost() const {
-    double totalCost = 0;
+Decimal VehicleAssembly::getCost() const {
+    Decimal totalCost = 0;
     for (const auto& part : parts) {
         totalCost += part->getCost();
     }
